02-data-analysis-tool/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
pandas
numpy
python-multipart
